+++ todo_app/app.py
@@ -4,19 +4,10 @@ from todo_app.data import trello_items
 
 from todo_app.flask_config import Config
 
+from todo_app.item import Item
+
 app = Flask(__name__)
 app.config.from_object(Config())
-
-class Item:
-    def __init__(self, id, name, status="To Do"):
-        self.id = id
-        self.name = name
-        self.status = status
-
-    @classmethod
-    def from_trello_card(cls, card, status):
-        return cls(card["id"], card["name"], status)
-
 
 @app.route("/")
 def index():
