+++ todo_app/item.py
@@ -1,0 +1,9 @@
+class Item:
+    def __init__(self, id, name, status="To Do"):
+        self.id = id
+        self.name = name
+        self.status = status
+
+    @classmethod
+    def from_trello_card(cls, card, status):
+        return cls(card["id"], card["name"], status)
